requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.1
contourpy==1.3.2
cycler==0.12.1
fonttools==4.57.0
idna==3.10
kiwisolver==1.4.8
labjack-ljm==1.23.0
matplotlib==3.10.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyvista==0.45.0
pyvistaqt==0.11.2
QtPy==2.4.3
requests==2.32.3
scipy==1.15.2
scooby==0.10.0
six==1.17.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vtk==9.4.2
